Add support for different treesitter langs
Prior to this change, there was no way to use a different treesitter
syntax for a given language.

This change adds a new field `treesitter_lang` to the language config
type and adds logic into the `treesitter.lua` file to ensure that the
language name will be overwritten with the custom treesitter langauge
name if specified.