Stop running Mason tools install on start
Prior to this change, mason tools install would run automatically on
start. This would cause an error for any tools that are awaiting
approval as a PR to the mason-registry repository.

This change makes the `mason-too-installer` only run manually for now.
This will stop the error being raised every time nvim is open.

> [!NOTE]
> You should manual run `MasonToolsInstall` of `MasonToolsUpdate` if
> you want the tools required for this config installed!